# ROX-28253: drop cursor in RunCursorQueryForSchemaFn