### Description

Cursor was introduced to prevent memory usage but in the meantime we improved how stores handles row and we don't need cursors at all and we can drop them and use normal queries instead.

- https://github.com/stackrox/stackrox/pull/2157
- https://github.com/jackc/pgx/discussions/1391

---

- [x] CHANGELOG update is not needed
- [x] Documentation is not needed
### Testing
- [x] inspected CI results
#### Automated testing
- [x] modified existing tests
- [x] contributed **no automated tests**
#### How I validated my change
CI