@janisz 2025-03-13
@dashrews78 That's a good idea I'll create a k6 test for that