@janisz 2025-03-13
I found a benchmark for ExportDeployments 
```
pkg: github.com/stackrox/rox/central/deployment/service
                                         │   old.txt   │              new.txt               │
                                         │   sec/op    │   sec/op     vs base               │
Service_Export/10/No_Query-8               253.5µ ± 1%   253.7µ ± 8%       ~ (p=1.000 n=10)
Service_Export/10/Query_10%_of_dataset-8   212.6µ ± 0%   213.5µ ± 1%  +0.42% (p=0.029 n=10)
Service_Export/10/Query_90%_of_dataset-8   429.8µ ± 0%   429.1µ ± 1%       ~ (p=0.579 n=10)
geomean                                    285.0µ        285.4µ       +0.12%

                                         │   old.txt    │               new.txt               │
                                         │     B/op     │     B/op      vs base               │
Service_Export/10/No_Query-8               118.4Ki ± 0%   118.1Ki ± 0%  -0.27% (p=0.000 n=10)
Service_Export/10/Query_10%_of_dataset-8   29.89Ki ± 0%   29.87Ki ± 0%  -0.09% (p=0.000 n=10)
Service_Export/10/Query_90%_of_dataset-8   125.4Ki ± 0%   125.1Ki ± 0%  -0.23% (p=0.000 n=10)
geomean                                    76.29Ki        76.14Ki       -0.19%

                                         │   old.txt   │               new.txt                │
                                         │  allocs/op  │  allocs/op   vs base                 │
Service_Export/10/No_Query-8               2.121k ± 0%   2.119k ± 0%  -0.09% (p=0.000 n=10)
Service_Export/10/Query_10%_of_dataset-8    466.0 ± 0%    466.0 ± 0%       ~ (p=1.000 n=10) ¹
Service_Export/10/Query_90%_of_dataset-8   2.152k ± 0%   2.151k ± 0%  -0.05% (p=0.000 n=10)
geomean                                    1.286k        1.285k       -0.05%
```
It's not super impressive