@janisz 2025-03-14
@dashrews78 @mtodor I updated a PR and dropped cursor