@janisz 2025-04-01
Closing as we need more data how this will affect the database. So far we increase batch size and removed internal batching so we should be good.
- #14727