@janisz 2025-04-10
```diff
+ grpc_req_duration..............: avg=122.87ms min=120.88ms med=122.77ms max=126.78ms p(90)=124.34ms p(95)=125.87ms
- + grpc_req_duration: avg=139.77ms min=37.38ms  med=58.54ms  max=421.75ms p(90)=380.58ms p(95)=390.98ms
```